+++ Teams/Communications/website/src/.vuepress/styles/screen.styl
@@ -1,0 +1,50 @@
+*
+  box-sizing border-box
+
+body
+  font 16px/1.5 var(--serif)
+  background #fff
+
+a
+  text-decoration none
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6
+  line-height 1.25
+  font-weight bold
+  letter-spacing -0.02em
+  margin 0
+  font-family var(--sans)
+  margin-bottom 1.5rem
+
+#site-markdown
+  h2,
+  h3,
+  h4,
+  h5,
+  h6
+    margin-top 3rem
+
+  h2
+    border-top 1px solid var(--bc)
+    padding-top 3rem
+
+  h3
+    font-weight 400
+
+  h4
+    font-family var(--serif)
+
+  p
+    margin 1.5em 0
+    &:last-child
+      margin-bottom 0
+
+  li
+    margin-bottom 0.5rem
+    p
+      margin 0
