--- Teams/Communications/website/src/.vuepress/styles/variables.css
@@ -1,35 +1,0 @@
-:root {
-  --tertiary: hsl(288, 67%, 55%);
-  --secondary: hsl(217, 100%, 70%);
-  --primary: hsl(233, 96%, 65%);
-  --link: hsl(233, 96%, 65%);
-  --hover: #46509f;
-  --bright: hsl(0, 100%, 0%);
-  --txt: hsla(0, 100%, 0%, 0.8);
-  --dim: hsla(0, 100%, 0%, 0.667);
-  --bc: hsla(233, 24%, 75%, 0.175);
-  --bc-dim: hsla(233, 24%, 75%, 0.0875);
-  --app-fg: hsl(220, 33%, 98%);
-  --app-bg: hsl(0, 0%, 100%);
-  --app-nav: hsl(233, 38%, 14%);
-
-  --success: hsl(120, 58%, 55%);
-  --success-bc: hsl(120, 58%, 41%);
-  --warning: hsl(35, 100%, 50%);
-  --warning-bc: hsl(35, 75%, 38%);
-  --danger: hsl(0, 100%, 55%);
-  --danger-bc: hsl(0, 100%, 41%);
-
-  --z-modal: 1000;
-  --z-app-header: 100;
-  --z-app-menu: 99;
-  --z-tool-bar: 98;
-  --z-list-item: 10;
-  --z-default: 1;
-  --z-zero: 0;
-  --z-below: -1;
-
-  --serif: "Mercury SSm A", "Mercury SSm B", "Georgia", "Times New Roman", serif;
-  --sans: "Gotham SSm A", "Gotham SSm B", "Helvetica Neue", Arial, Helvetica,
-    sans-serif;
-}
